LayoutTests/fast/js/sputnik/Conformance/12_Statement/12.7_The_continue_Statement/S12.7_A1_T1-expected.txt

CONSOLE MESSAGE: line 80: SyntaxError: Parse error
S12.7_A1_T1

PASS Expected parsing failure

TEST COMPLETE
